Saptamana 4 + integrare FE

+++ fiipractic/src/main/java/com/fiipractic/Repository/RequestHistoryRepository.java
@@ -7,6 +7,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RequestHistoryRepository extends JpaRepository<RequestHistory, Integer> {
+public interface RequestHistoryRepository extends JpaRepository<RequestHistory, Long> {
     Page<RequestHistory> findByUserId(Long userId, Pageable pageable);
 }
